fix: add crossorigin to audio and video elements
