+++ packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/attribute-crossorigin/actual.html
@@ -1,0 +1,6 @@
+<template>
+    <!-- Issue #917 -->
+    <img src="http://www.example.com/image.png" crossorigin="anonymous">
+    <video src="http://www.example.com/video.mp4" crossorigin="anonymous"></video>
+    <audio src="http://www.example.com/video.mp3" crossorigin="anonymous"></audio>
+</template>
