+++ packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/attribute-crossorigin/metadata.json
