--- packages/@lwc/template-compiler/src/__tests__/fixtures/attributes/img/actual.html
@@ -1,4 +1,0 @@
-<template>
-    <!-- Issue #917 -->
-    <img src="http://www.example.com/image.png" crossorigin="anonymous">
-</template>
